+++ halotools/mock_observables/isolation_functions/isolation_functions_helpers.py
